Moved travels to separate file
