+++ src/travels.ts
@@ -1,0 +1,17 @@
+export interface Travel {
+  countryCode: string;
+}
+
+const travels = [
+  {
+    countryCode: "ZM",
+  },
+  {
+    countryCode: "TW",
+  },
+  {
+    countryCode: "SK",
+  },
+];
+
+export default travels;
